feat(generic): abstract diesel and solar for Cooking and Facilities
